Remove CS350/550
Since the students of CS350/550 are sharing the same course webpage and
assignments, by default, there is no need to specify CS350/550 in the
assignments. If a document is specific to CS350 or CS550, we can specify it on
the webpage and in the document.

modified:   shell_fg_bg.html